# docs/swagger.yaml
definitions:
  dao.UpdateSite:
    properties:
      address:
        type: string
      all_open:
        description: 指针类型，可以确保使用Updates方法更新时，如果值为false时也能更新成功
        type: boolean
      callback_url:
        description: 指针类型，可以确保使用Updates方法更新时，如果值为空时也能更新成功
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sso:
        description: 指针类型，可以确保使用Updates方法更新时，如果值为false时也能更新成功
        type: boolean
      sso_type:
        type: integer
    type: object
...
  service.Authorize:
    properties:
      client_id:
        type: string
      redirect_uri:
        type: string
      response_type:
        type: string
      scope:
        type: string
      state:
        type: string
    required:
    - client_id
    - response_type
    type: object
...
  service.ResponseToken:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
  service.ResponseUserinfo:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
...
  service.SiteCreate:
    properties:
      address:
        type: string
      callback_url:
        type: string
      description:
        type: string
      icon:
        type: string
      name:
        type: string
      site_group_id:
        type: integer
      sso:
        type: boolean
      sso_type:
        type: integer
    required:
    - address
    - description
    - name
    - site_group_id
    - sso
    type: object
  service.SiteGroupCreate:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  service.SiteGroupUpdate:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  service.SiteUserUpdate:
    properties:
      id:
        type: integer
      users:
        items:
          type: integer
        type: array
    required:
    - id
    - users
    type: object
  service.Token:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      code:
        type: string
      grant_type:
        type: string
      redirect_uri:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/oauth/authorize:
    post:
      description: OAuth2.0认证相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 授权请求参数
        in: body
        name: authorize
        required: true
        schema:
          $ref: '#/definitions/service.Authorize'
      responses:
        "200":
          description: '{"code": 0, "msg": 授权成功, "redirect_uri": redirect_uri}'
          schema:
            type: string
      summary: 客户端授权
      tags:
      - OAuth2.0认证
  /api/v1/oauth/token:
    post:
      description: OAuth2.0认证相关接口
      parameters:
      - description: 授权请求参数
        in: body
        name: authorize
        required: true
        schema:
          $ref: '#/definitions/service.Token'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseToken'
      summary: 客户端认证
      tags:
      - OAuth2.0认证
  /api/v1/oauth/userinfo:
    get:
      description: OAuth2.0认证相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseUserinfo'
      summary: 获取用户信息
      tags:
      - OAuth2.0认证
...
  /api/v1/site:
    post:
      consumes:
      - application/json
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分组信息
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/service.SiteCreate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0, "msg": "创建成功", "data": nil}'
          schema:
            type: string
      summary: 创建站点
      tags:
      - 站点管理
    put:
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分组信息
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/dao.UpdateSite'
      responses:
        "200":
          description: '{"code": 0, "msg": "更新成功", "data": nil}'
          schema:
            type: string
      summary: 更新站点信息
      tags:
      - 站点管理
  /api/v1/site/{id}:
    delete:
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分组ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 0, "msg": "删除成功", "data": nil}'
          schema:
            type: string
      summary: 删除站点
      tags:
      - 站点管理
  /api/v1/site/group:
    post:
      consumes:
      - application/json
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分组信息
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/service.SiteGroupCreate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0, "msg": "创建成功", "data": nil}'
          schema:
            type: string
      summary: 创建分组
      tags:
      - 站点管理
    put:
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分组信息
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/service.SiteGroupUpdate'
      responses:
        "200":
          description: '{"code": 0, "msg": "更新成功", "data": nil}'
          schema:
            type: string
      summary: 更新分组信息
      tags:
      - 站点管理
  /api/v1/site/group/{id}:
    delete:
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分组ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 0, "msg": "删除成功", "data": nil}'
          schema:
            type: string
      summary: 删除站点分组
      tags:
      - 站点管理
  /api/v1/site/guide:
    get:
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: '{"code": 0, "data": []}'
          schema:
            type: string
      summary: 获取站点列表
      tags:
      - 站点管理
  /api/v1/site/logoUpload:
    post:
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 头像
        in: formData
        name: logo
        required: true
        type: file
      responses:
        "200":
          description: '{"code": 0, "path": logoPath}'
          schema:
            type: string
      summary: 站点图片上传
      tags:
      - 站点管理
  /api/v1/site/users:
    put:
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户信息
        in: body
        name: users
        required: true
        schema:
          $ref: '#/definitions/service.SiteUserUpdate'
      responses:
        "200":
          description: '{"code": 0, "msg": "更新成功", "data": nil}'
          schema:
            type: string
      summary: 更新站点用户
      tags:
      - 站点管理
  /api/v1/sites:
    get:
      description: 站点关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分页
        in: query
        name: page
        required: true
        type: integer
      - description: 分页大小
        in: query
        name: limit
        required: true
        type: integer
      - description: 站点名称
        in: query
        name: name
        type: string
      responses:
        "200":
          description: '{"code": 0, "data": []}'
          schema:
            type: string
      summary: 获取站点列表
      tags:
      - 站点管理
...
swagger: "2.0"